フォントのレンダリングを一旦画像に行う仕様に変更して軽量化

+++ CommentPrompter/src/monolith52/comprompt/animation/Animation.java
@@ -4,7 +4,7 @@ public interface Animation {
 	public void step();
 	public int getX();
 	public int getY();
-	public int getAlpha();
+	public float getAlpha();
 	public void setOnFinish(Runnable onFinish);
 	public boolean isFinished();
 }
